requirements.txt
python-valve
flask
jinja2
requests
